user-service/backend/requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
mysql-connector-python==9.0.0
python-dotenv==1.0.0
Werkzeug>=2.3.3